Fix webhook flow bugs (#21)
* Fix webhook flow bugs

* Change pool_entries.value to NUMERIC(50, 0)

* Fix broken tests & remove unused code

* Update src/webhookApi/handlers/poolInteractionsHandler.ts


* Remove explicit value::numeric conversion

---------

+++ resources/db/migrations/0000_lyrical_ser_duncan.sql
@@ -10,7 +10,7 @@ CREATE TABLE IF NOT EXISTS "pool_entries" (
 	"entrant" varchar(44) NOT NULL,
 	"option" varchar(44) NOT NULL,
 	"pool" varchar(44) NOT NULL,
-	"value" bigint NOT NULL
+	"value" numeric(50, 0) NOT NULL
 );
 --> statement-breakpoint
 DO $$ BEGIN
